requirements.txt
asgiref==3.11.0
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.0.0
psycopg==3.3.2
psycopg-binary==3.3.2
PyJWT==2.10.1
sqlparse==0.5.4
typing_extensions==4.15.0
